[QOL-5834] gracefully handle missing DataRequests or Scheming extensions
- this will enable us to more easily reuse this plugin for the Publications theme
