+++ ckanext/data_qld/plugin.py
@@ -4,15 +4,28 @@ import cgi
 import ckan.plugins as plugins
 import ckan.plugins.toolkit as toolkit
 
-import actions
 import auth_functions as auth
 import constants
 import converters
 import helpers
-import validation
 import logging
 
 log = logging.getLogger(__name__)
+
+try:
+    import actions
+    import datarequest_auth_functions as datareq_auth
+    enable_datarequests = True
+except ImportError:
+    log.warn("Unable to find data request requirements, disabling")
+    enable_datarequests = False
+
+try:
+    import validation
+    enable_validation = True
+except ImportError:
+    log.warn("Unable to find validation requirements, disabling")
+    enable_validation = False
 
 
 class DataQldPlugin(plugins.SingletonPlugin):
@@ -56,11 +69,15 @@ class DataQldPlugin(plugins.SingletonPlugin):
 
     # IValidators
     def get_validators(self):
-        return {
+        validators = {
             'data_qld_filesize_converter': converters.filesize_converter,
             'data_qld_filesize_formatter': converters.filesize_formatter,
-            'data_qld_scheming_choices': validation.scheming_choices,
         }
+
+        if enable_validation:
+            validators['data_qld_scheming_choices'] = validation.scheming_choices
+
+        return validators
 
     # IPackageController
     def set_maintainer_from_author(self, entity):
@@ -76,17 +93,21 @@ class DataQldPlugin(plugins.SingletonPlugin):
 
     # IAuthFunctions
     def get_auth_functions(self):
-        auth_functions = {
-            constants.UPDATE_DATAREQUEST: auth.update_datarequest,
-            constants.UPDATE_DATAREQUEST_ORGANISATION: auth.update_datarequest_organisation,
-            constants.CLOSE_DATAREQUEST: auth.close_datarequest,
-            constants.OPEN_DATAREQUEST: auth.open_datarequest,
-            'member_create': auth.member_create
-        }
+        auth_functions = {'member_create': auth.member_create}
+
+        if enable_datarequests:
+            auth_functions[constants.UPDATE_DATAREQUEST] = datareq_auth.update_datarequest
+            auth_functions[constants.UPDATE_DATAREQUEST_ORGANISATION] = datareq_auth.update_datarequest_organisation
+            auth_functions[constants.CLOSE_DATAREQUEST] = datareq_auth.close_datarequest
+            auth_functions[constants.OPEN_DATAREQUEST] = datareq_auth.open_datarequest
+
         return auth_functions
 
     # IActions
     def get_actions(self):
+        if not enable_datarequests:
+            return {}
+
         additional_actions = {
             constants.OPEN_DATAREQUEST: actions.open_datarequest,
             constants.CREATE_DATAREQUEST: actions.create_datarequest,
@@ -97,14 +118,16 @@ class DataQldPlugin(plugins.SingletonPlugin):
 
     # IRoutes
     def before_map(self, m):
-        # Re_Open a Data Request
-        m.connect('/%s/open/{id}' % constants.DATAREQUESTS_MAIN_PATH,
-                  controller='ckanext.data_qld.controller:DataQldUI',
-                  action='open_datarequest', conditions=dict(method=['GET', 'POST']))
+        if enable_datarequests:
+            # Re_Open a Data Request
+            m.connect('/%s/open/{id}' % constants.DATAREQUESTS_MAIN_PATH,
+                      controller='ckanext.data_qld.controller:DataQldUI',
+                      action='open_datarequest', conditions=dict(method=['GET', 'POST']))
 
-        m.connect('/dataset/{dataset_id}/resource/{resource_id}/%s/show/' % constants.SCHEMA_MAIN_PATH,
-                  controller='ckanext.data_qld.controller:DataQldUI',
-                  action='show_schema', conditions=dict(method=['GET']))
+        if enable_validation:
+            m.connect('/dataset/{dataset_id}/resource/{resource_id}/%s/show/' % constants.SCHEMA_MAIN_PATH,
+                      controller='ckanext.data_qld.controller:DataQldUI',
+                      action='show_schema', conditions=dict(method=['GET']))
 
         return m
